# swagger/swagger.yaml
basePath: /
definitions:
...
  server.APIStatus:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  server.Event:
    properties:
      action_name:
        type: string
      client_ip:
        type: string
      client_user_agent:
        type: string
      created_at:
        type: string
      geo_ip_location:
        type: string
      http_endpoint:
        type: string
      http_method:
        type: string
      http_status:
        type: string
      id:
        type: string
      organization_id:
        type: string
      type:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  server.createEventRequest:
    properties:
      action_name:
        type: string
      client_ip:
        type: string
      client_user_agent:
        type: string
      geo_ip_location:
        type: string
      http_endpoint:
        type: string
      http_method:
        type: string
      http_status:
        type: string
      type:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  server.fetchEventResponse:
    properties:
      event:
        $ref: '#/definitions/server.Event'
      message:
        type: string
    required:
    - message
    type: object
  server.fetchEventsResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/server.Event'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/server.meta'
    required:
    - message
    type: object
...
  server.meta:
    properties:
      paging:
        $ref: '#/definitions/server.pagingInfo'
    required:
    - paging
    type: object
  server.pagingInfo:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    required:
    - page
    - per_page
    - total
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Logtrace API documentation
  title: Logtrace API
  version: 0.1.0
paths:
...
  /v1/events:
    get:
      consumes:
      - application/json
      description: Get a list of events with optional filters and pagination
      parameters:
      - description: Filter by HTTP status
        in: query
        name: http_status
        type: string
      - description: Filter by HTTP method
        in: query
        name: http_method
        type: string
      - description: Filter by start date (ISO format)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (ISO format)
        in: query
        name: end_date
        type: string
      - description: Filter by username
        in: query
        name: username
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Search term for action name or endpoint
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Events fetched successfully
          schema:
            $ref: '#/definitions/server.fetchEventsResponse'
        "500":
          description: Failed to fetch events
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Create a new event log entry
      parameters:
      - description: Event creation request
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/server.createEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Event created successfully
          schema:
            $ref: '#/definitions/server.APIStatus'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Could not create event at this time. an error occurred
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Events
  /v1/events/{reference}:
    get:
      consumes:
      - application/json
      description: Get a specific event by ID
      parameters:
      - description: Event ID
        in: path
        name: reference
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event has been fetched successfully
          schema:
            $ref: '#/definitions/server.fetchEventResponse'
        "400":
          description: Invalid event reference
          schema:
            $ref: '#/definitions/server.APIStatus'
        "404":
          description: Event not found
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Failed to fetch event
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Events
...
schemes:
- http
- https
swagger: "2.0"